Fix "Break After First Match" when the same file is open multiple times
"Break After First Match" broke tab coloring when the same file was open
in multiple tabs. When a tab was colored with a regex rule, the file's
path was saved to a list, so that a later color rule wouldn't override
it.

However, saving the path doesn't really work, because the same path can
be open in multiple tabs, which would break coloring for all tabs with
that file path. The current code just breaks out of the loop after a
color rule match, which seems to work, unless I'm missing an edge case.